+++ server/api/src/main/java/com/walmartlabs/concord/server/api/user/UserType.java
@@ -1,0 +1,7 @@
+package com.walmartlabs.concord.server.api.user;
+
+public enum UserType {
+
+    LOCAL,
+    LDAP
+}