+++ server/impl/src/test/java/com/walmartlabs/concord/server/user/UserDaoTest.java
@@ -22,6 +22,7 @@ package com.walmartlabs.concord.server.user;
 
 import com.walmartlabs.concord.server.AbstractDaoTest;
 import com.walmartlabs.concord.server.SecureRandomProvider;
+import com.walmartlabs.concord.server.api.user.UserType;
 import com.walmartlabs.concord.server.security.apikey.ApiKeyDao;
 import org.junit.Before;
 import org.junit.Ignore;
@@ -51,7 +52,7 @@ public class UserDaoTest extends AbstractDaoTest {
         String username = "user#" + System.currentTimeMillis();
         Set<String> permissions = Collections.singleton("*");
 
-        UUID userId = userDao.insert(username, permissions, false);
+        UUID userId = userDao.insert(username, UserType.LOCAL, permissions, false);
 
         String s = "key#" + System.currentTimeMillis();
         String apiKey = Base64.getEncoder().encodeToString(s.getBytes());